Prices are now encoded as a Money instance
This makes it easier to calculate rounded values and such

+++ src/ui/nina/positioncontainermodel.cpp
@@ -159,6 +159,6 @@ PositionContainerModel::set(const nina::domain::PositionContainer& positions)
         append(QString::fromStdString(position.getDescription()),
                QString::fromStdString(position.getUnit()),
                position.getNumUnits(),
-               position.getPricePerUnit()
+               position.getPricePerUnit().toDouble()
             );
 }
